--- a/networkx/algorithms/__init__.py
+++ b/networkx/algorithms/__init__.py
@@ -51,6 +51,7 @@ import networkx.algorithms.traversal
 import networkx.algorithms.chordal
 import networkx.algorithms.operators
 import networkx.algorithms.tree
+import networkx.algorithms.bidirectional
 
 # bipartite
 from networkx.algorithms.bipartite import (projected_graph, project, is_bipartite,