--- a/networkx/algorithms/bidirectional.py
+++ b/networkx/algorithms/bidirectional.py
@@ -6,7 +6,9 @@
 import networkx as nx
 __author__ = """\n""".join(['André Dietrich <dietrich@ivs.cs.uni-magdeburg.de>',
                             'Sebastian Zug <zug@ivs.cs.uni-magdeburg.de>'])
-__all__ = ['all_simple_paths_ordered']
+__all__ = ['all_simple_paths',
+	   'all_shortest_paths',
+	   'has_path']
 
 
 def all_simple_paths(G, source, target, cutoff=None):
@@ -43,7 +45,7 @@ def all_simple_paths(G, source, target, cutoff=None):
     >>> import networkx as nx
     >>>
     >>> G = nx.complete_graph(4)
-    >>> for path in all_simple_paths(G, source=0, target=3):
+    >>> for path in nx.bidirectional.all_simple_paths(G, source=0, target=3):
     ...     print(path)
     ...
     [0, 3]
@@ -51,7 +53,7 @@ def all_simple_paths(G, source, target, cutoff=None):
     [0, 2, 3]
     [0, 2, 1, 3]
     [0, 1, 2, 3]
-    >>> paths = all_simple_paths(G, source=0, target=3, cutoff=2)
+    >>> paths = nx.bidirectional.all_simple_paths(G, source=0, target=3, cutoff=2)
     >>> print(list(paths))
     [[0, 3], [0, 1, 3], [0, 2, 3]]
 
@@ -93,7 +95,8 @@ def all_simple_paths(G, source, target, cutoff=None):
 
     See Also
     --------
-    nx.all_simple_paths, nx.all_shortest_paths, nx.shortest_path
+    nx.all_simple_paths, nx.all_shortest_paths, nx.shortest_path,
+    nx.bidirectional.all_shortest_paths, nx.bidirectional.has_path
     """
     
     if source not in G:
@@ -191,7 +194,7 @@ def all_shortest_paths(G, source, target):
     >>> import networkx as nx
     >>>
     >>> G = nx.complete_graph(4)
-    >>> for path in all_shortest_paths(G, source=0, target=3):
+    >>> for path in nx.bidirectional.all_shortest_paths(G, source=0, target=3):
     ...     print(path)
     ...
     [0, 3]
@@ -234,7 +237,8 @@ def all_shortest_paths(G, source, target):
 
     See Also
     --------
-    nx.all_simple_paths, nx.all_shortest_paths, nx.shortest_path
+    nx.all_simple_paths, nx.all_shortest_paths, nx.shortest_path,
+    nx.bidirectional.all_simple_paths, nx.bidirectional.has_path
     """