… networkx/algorithms/tests/test_simple_paths.py