@OrkoHunter 2015-06-24
Very nice effort @andre-dietrich. However I personally think creating a wiki or some other page was a good idea for all the benchmarks of your code. Loading this junk every time a new comment is created, sometimes feels weird. Anyways, your call.