@MridulS 2015-06-24
Please do add tests for this, and try to format your code according to PEP8 :)