@andre-dietrich 2015-06-25
Hi, everything should now also be formatted according to PEP8, I also included additional tests (whereby I reused already existing ones). The tests were actually a good hint, because I missed directed graphs, which are now also supported. Thus, there are now three additional functions.