@chebee7i 2015-06-25
We can iterate on style endlessly :) Has anyone had a chance to review the substance? Unfortunately, it will be a few days before I can even attempt that.