@OrkoHunter 2015-06-25
@chebee7i For smaller graphs (I checked for around 1000), `nx.bidirectional.all_simple_paths()` takes little bit more time than the original `nx.all_simple_paths()`. The benchmarks claims that this is other way round for very large number of nodes.