@andre-dietrich 2015-06-25
@OrkoHunter can you post your graph? Is it a fully connected graph? What was the cutoff number you applied?