@OrkoHunter 2015-06-25
This is pretty much what I tried.
```
In [20]: G = nx.complete_graph(1000)

In [21]: %%timeit
   ....: nx.all_simple_paths(G, 1, 50)
   ....: 
The slowest run took 7.49 times longer than the fastest. This could mean that an intermediate result is being cached 
1000000 loops, best of 3: 1.47 µs per loop

In [22]: %%timeit
   ....: nx.bidirectional.all_simple_paths(G, 1, 50)
   ....: 
The slowest run took 7.39 times longer than the fastest. This could mean that an intermediate result is being cached 
1000000 loops, best of 3: 1.64 µs per loop
```