@andre-dietrich 2015-06-25
For fully connected graphs we are slower, but I mentioned it in the notes of the function description... The less connected the graph, the better we perform. Otherwise we generate to much overhead...