@jtorrents 2015-06-27
Very nice indeed! Thanks for working on this @andre-dietrich . I did some quick speed tests with sparse graphs with right skewed degree distributions (a pretty usual use case), and it seems that the proposed implementation is much slower than `nx.all_simple_paths`. See: 

```python
In [135]: %paste
import networkx as nx
from networkx.utils import powerlaw_sequence, create_degree_sequence

def pick_nodes(G):
    u = next(G.nodes_iter())
    v = next(nx.non_neighbors(G, u))
    return u, v

def build_power_law(n):
    deg_seq = create_degree_sequence(n, powerlaw_sequence, 100)
    G = nx.Graph(nx.configuration_model(deg_seq))
    G.remove_edges_from(G.selfloop_edges())
    G = max(nx.biconnected_component_subgraphs(G), key=len)
    G.name = 'Biconnected component of power law configuration model: {0}'.format(n)
    return G
## -- End pasted text --

In [136]: G = build_power_law(40)

In [137]: print(nx.info(G))
Name: Biconnected component of power law configuration model: 40
Type: Graph
Number of nodes: 22
Number of edges: 44
Average degree:   4.0000

In [138]: u, v = pick_nodes(G)

In [139]: timeit paths = list(nx.all_simple_paths(G, u, v))
10 loops, best of 3: 27.4 ms per loop

In [140]: timeit paths = list(nx.shortest_simple_paths(G, u, v))
1 loops, best of 3: 2.66 s per loop

In [141]: timeit paths = list(nx.bidirectional.all_simple_paths(G, u, v))
1 loops, best of 3: 30.6 s per loop
```

@OrkoHunter note that in your speed tests your are not actually computing the paths, you are only creating the generator.

@andre-dietrich  note that we also have a recently added `nx.shortest_simple_paths` which is based on Yen's algorithm for computing k shortest simple paths.