@andre-dietrich 2015-06-27
@jtorrents Thanks for reviewing ;)

I guess, all simple paths is a bit of problematic in our approach, take a look at the last tow diagrams in our approach. There is some kind of break even point, we are better until a certain cutoff number is reached, after that number we perform worse... That is indicated by the two green curved lines, while the blue line appears to be linear (well in logarithmic scale)... if you test again with some cutoff values, we should perform better, see the code below. 

I took your code with a bit larger graph. The last three lines, show the all_shortest_path approaches. In our solution to this, we could even reduce the effort with a slightly adopted version of our algorithm. Unfortunately, I had to stop your mentioned "nx.shortest_simple_paths", because it took longer than two minutes ...

Cheers,

André

In [2]: G = build_power_law(100)

In [3]: print(nx.info(G))
Name: Biconnected component of power law configuration model: 100
Type: Graph
Number of nodes: 59
Number of edges: 156
Average degree:   5.2881

In [4]: u, v = pick_nodes(G)

In [5]: print len(list(nx.all_simple_paths(G, u, v, cutoff=10)))
278959

In [7]: timeit paths = list(nx.all_simple_paths(G, u, v, cutoff=10))
1 loops, best of 3: 27.2 s per loop

In [8]: timeit paths = list(nx.bidirectional.all_simple_paths(G, u, v, cutoff=10))
1 loops, best of 3: 6.25 s per loop

In [9]: timeit paths = list(nx.all_shortest_paths(G, u, v))
10000 loops, best of 3: 75.5 µs per loop

In [10]: timeit paths = list(nx.bidirectional.all_shortest_paths(G, u, v))
The slowest run took 5.56 times longer than the fastest. This could mean that an intermediate result is being cached 
100000 loops, best of 3: 12.4 µs per loop

In [*]: timeit paths = list(nx.shortest_simple_paths(G, u, v))