@andre-dietrich 2015-06-29
@jtorrents PS: nx.bidirectional.all_simple_paths(G, u, v, cutoff=10)) returns the results ordered, according to their path length, which from my point of view is a nice feature ;)